+++ kotti_settings/tests/test_util.py
@@ -25,3 +25,28 @@ def test_add_settings_with_missing_attributes(db_session):
 
     TestSettings['settings'][0]['title'] = 'test title'
     assert add_settings(TestSettings) == None
+
+
+def test_get_setting(db_session, root):
+    from kotti_settings.util import add_settings
+    from kotti_settings.util import get_setting
+    test_settings = {
+        'name': 'test-get-setting',
+        'title': "Some test settings.",
+        'settings': [
+            {'type': 'String',
+             'name': 'first_test_setting',
+             'title': 'first test setting',
+             'default': 'first test string', },
+            {'type': 'Integer',
+             'name': 'second_test_setting',
+             'title': 'second test setting',
+             'default': 5, },
+        ]
+    }
+    add_settings(test_settings)
+
+    first = get_setting('first_test_setting')
+    assert first == 'first test string'
+    second = get_setting('test_util-second_test_setting')
+    assert second == 5