+++ CHANGES.rst
@@ -4,6 +4,8 @@ Changelog
 0.1a2(Unreleased)
 -----------------
 
+* Added util method get_setting as default way to get a setting.
+
 
 0.1a1(2013-03-06)
 -----------------